pyproject.toml
[tool.poetry]
name = "StanTasq"
version = "0.1.0"
description = ""
authors = ["Adam Ryczkowski <adam@statystyka.net>"]
readme = "README.adoc"

[tool.poetry.dependencies]
python = "^3.12"
taskiq = { version = "^0.11.10", optional = false, extras = ["reload"] }
loguru = "^0.7.3"
taskiq-dependencies = "^1.5.6"
anyio = "^4.7.0"
taskiq-nats = "^0.5.1"
cmdstanpy = "^1.2.5"
overrides = "^7.7.0"
numpy = "^2.2.1"
valuewitherror = {path = "/home/adam/Documents/praca/TrainerEngine/lib/ValueWithError"}#, version = "^2.4.0"}
cachemanager = {path = "/home/adam/Documents/praca/TrainerEngine/lib/CacheManager"}#, version = "^2.4.0"}
humanize = "^4.11.0"
prettytable = "^3.12.0"
tqdm = "^4.67.1"
pytest-asyncio = "^0.25.1"
pydantic = "^2.10.4"
numpydantic = "^1.6.7"

[poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.poetry.scripts]
#run-server = "mikapi:app.run"

[tool.pytest.ini_options]
asyncio_mode = "auto"
